+++ packages/effects-webgpu/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @glissade/effects-webgpu
 
+## 0.63.2-pre.0
+
+### Patch Changes
+
+- @glissade/backend-canvas2d@0.63.2-pre.0
+- @glissade/scene@0.63.2-pre.0
+
 ## 0.63.1
 
 ### Patch Changes