+++ packages/sfx/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @glissade/sfx
 
+## 0.63.2-pre.0
+
+### Patch Changes
+
+- @glissade/core@0.63.2-pre.0
+
 ## 0.63.1
 
 ### Patch Changes